+++ @remirror/core/src/__tests__/extension-manager.spec.tsx
@@ -6,12 +6,12 @@ import { Extension } from '../extension';
 import { ExtensionManager, isExtensionManager } from '../extension-manager';
 import { Cast } from '../helpers/base';
 import { NodeExtension } from '../node-extension';
-import { NodeViewPortalContainer } from '../portal-container';
+import { PortalContainer } from '../portal-container';
 import { Attrs, EditorState, NodeExtensionSpec } from '../types';
 
 export const helpers = {
   getState: Cast(jest.fn(() => state)),
-  portalContainer: new NodeViewPortalContainer(),
+  portalContainer: new PortalContainer(),
 };
 
 const innerMock = jest.fn();