+++ @remirror/react-ssr/src/__tests__/remirror-ssr.server.spec.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { createTestManager, initialJson } from '@test-fixtures/schema-helpers';
 import { renderToString } from 'react-dom/server';
 
-import { ExtensionManager, NodeViewPortalContainer } from '@remirror/core';
+import { ExtensionManager, PortalContainer } from '@remirror/core';
 import { RemirrorSSR } from '..';
 
 let manager: ExtensionManager;
@@ -17,7 +17,7 @@ beforeEach(() => {
 });
 
 test('should render the ssr component', () => {
-  manager.init({ getState: () => state, portalContainer: new NodeViewPortalContainer() });
+  manager.init({ getState: () => state, portalContainer: new PortalContainer() });
   const state = manager.createState({ content: initialJson });
 
   const htmlString = renderToString(