+++ @remirror/react-ssr/src/__tests__/remirror-ssr.spec.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { createTestManager, initialJson } from '@test-fixtures/schema-helpers';
 import { render } from '@testing-library/react';
 
-import { ExtensionManager, NodeViewPortalContainer } from '@remirror/core';
+import { ExtensionManager, PortalContainer } from '@remirror/core';
 import { RemirrorSSR } from '..';
 
 let manager: ExtensionManager;
@@ -13,7 +13,7 @@ beforeEach(() => {
 });
 
 test('should render the ssr component', () => {
-  manager.init({ getState: () => state, portalContainer: new NodeViewPortalContainer() });
+  manager.init({ getState: () => state, portalContainer: new PortalContainer() });
   const state = manager.createState({ content: initialJson });
 
   const { container } = render(