+++ @remirror/react/src/components/__tests__/remirror-portals.spec.tsx
@@ -1,11 +1,11 @@
-import { NodeViewPortalContainer } from '@remirror/core';
+import { PortalContainer } from '@remirror/core';
 import { act, render } from '@testing-library/react';
-import { NodeViewPortalComponent } from '../node-view-portal';
+import { RemirrorPortals } from '../remirror-portals';
 
 describe('NodeViewPortalComponent', () => {
   it('should update on render', () => {
-    const portalContainer = new NodeViewPortalContainer();
-    const { getByTestId } = render(<NodeViewPortalComponent portalContainer={portalContainer} />);
+    const portalContainer = new PortalContainer();
+    const { getByTestId } = render(<RemirrorPortals portalContainer={portalContainer} />);
 
     act(() => {
       const mockRender = jest.fn(() => <div data-testid='test' />);