+++ @remirror/react/src/components/index.ts
@@ -2,4 +2,4 @@ export * from './remirror';
 export * from './remirror-manager';
 export * from './remirror-extension';
 export * from './providers';
-export * from './node-view-portal';
+export * from './remirror-portals';